requirements.txt
alembic==1.3.1
aniso8601==6.0.0
astroid==2.2.5
attrs==19.3.0
autopep8==1.4.4
awscli==1.16.309
Babel==2.7.0
botocore==1.13.45
cffi==1.13.2
Click==7.0
colorama==0.4.1
cryptography==2.8
docutils==0.15.2
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
gunicorn==20.0.4
importlib-metadata==1.3.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jwt==0.6.1
lazy-object-proxy==1.4.0
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pep8==1.7.1
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.2
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.6
pytest==5.3.2
python-dateutil==2.6.0
python-dotenv==0.10.3
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.2
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.8
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
zipp==0.6.0
